# docs/openapi/oms.yaml
openapi: 3.0.3
info:
  title: Shop AG Order Management API
  description: |
    REST API des Order Management Systems (OMS) zur Erfassung und Nachverfolgung von Aufträgen.
  version: "1.0.0"
servers:
  - url: http://localhost:8000
    description: Lokale Entwicklungsumgebung
tags:
  - name: Orders
    description: Verwaltung von Kundenbestellungen
paths:
  /orders:
    post:
      tags:
        - Orders
      summary: Bestellung erfassen
      description: |
        Erfasst eine neue Kundenbestellung und startet den Integrationsworkflow.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "201":
          description: Bestellung wurde entgegengenommen.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Ungültige Bestelldaten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Interner Fehler bei der Verarbeitung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Bestellung abrufen
      description: Liefert den aktuellen Status einer Bestellung inklusive Prozesshistorie.
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Details und Status der Bestellung.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "404":
          description: Bestellung nicht gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    OrderCreateRequest:
      type: object
      required:
        - orderId
        - customer
        - items
        - totalAmount
        - shippingAddress
      properties:
        orderId:
          type: string
          example: ORD-2025-10-16-7891
        customer:
          $ref: "#/components/schemas/Customer"
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
          format: float
          example: 209.97
        shippingAddress:
          $ref: "#/components/schemas/Address"
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Optionale Zusatzinformationen (z.B. für Simulationen).
    OrderResponse:
      type: object
      required:
        - orderId
        - status
        - createdAt
        - history
      properties:
        orderId:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        history:
          type: array
          description: Chronologisch sortierte Statusmeldungen zum Auftrag.
          items:
            $ref: "#/components/schemas/StatusEntry"
    Customer:
      type: object
      required:
        - customerId
        - prename
        - name
      properties:
        customerId:
          type: string
          example: CUST-45823
        prename:
          type: string
          example: Andreas
        name:
          type: string
          example: Heberle
    OrderItem:
      type: object
      required:
        - productId
        - quantity
        - price
      properties:
        productId:
          type: string
          example: P-8821
        quantity:
          type: integer
          minimum: 1
          example: 2
        price:
          type: number
          format: float
          example: 29.99
    Address:
      type: object
      required:
        - street
        - city
        - zipCode
        - country
      properties:
        street:
          type: string
          example: Moltke Str. 30
        city:
          type: string
          example: Karlsruhe
        zipCode:
          type: string
          example: "76133"
        country:
          type: string
          example: Germany
    StatusEntry:
      type: object
      required:
        - timestamp
        - status
        - message
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/OrderStatus"
        message:
          type: string
    OrderStatus:
      type: string
      enum:
        - RECEIVED
        - INVENTORY_RESERVED
        - INVENTORY_UNAVAILABLE
        - PAYMENT_AUTHORIZED
        - PAYMENT_FAILED
        - FULFILLMENT_STARTED
        - ITEMS_PICKED
        - ORDER_PACKED
        - ORDER_SHIPPED
        - CANCELLED
        - FAILED
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string